Backing out of requesting alt_identifiers every time.
This is supposed to address bug #522.  The code as it was up to here
would have needed aggregation of alt_identifiers (which are n:1 over
resources), or else we see duplicate capabilities.

But at least some registry operators prefer to not hit the rr.alt_identifier
table by default as long as it's not clear who will actually look at these
alternate identifiers.

But we maintain the alt identifiers in describe(); to do that, there's now
get_alternate_identifiers method returning these.

The downside: describe() now does an uncached network query.  Perhaps
we want to at least hide failures from there?  On the other hand, once
we are here we can also call get_contact() here; should we?